# backend/public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Warehouse Service Management API",
        "description": "API for managing warehouse operations.",
        "contact": {
            "email": "support@yourcompany.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/category": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Lấy danh sách tất cả các danh mục",
                "description": "Trả về danh sách tất cả các danh mục trong hệ thống.",
                "operationId": "getAllCategories",
                "responses": {
                    "200": {
                        "description": "Lấy danh sách danh mục thành công."
                    },
                    "500": {
                        "description": "Lỗi máy chủ nội bộ."
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Tạo một danh mục mới",
                "description": "Thêm mới một danh mục",
                "operationId": "addCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_name",
                                    "category_type"
                                ],
                                "properties": {
                                    "category_name": {
                                        "type": "string",
                                        "example": "Danh mục A"
                                    },
                                    "category_type": {
                                        "type": "string",
                                        "enum": [
                                            "product",
                                            "service",
                                            "post"
                                        ],
                                        "example": "product"
                                    },
                                    "slugs": {
                                        "type": "string",
                                        "example": "danh-muc-a"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Mô tả về danh mục A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Danh mục được tạo thành công"
                    },
                    "400": {
                        "description": "Dữ liệu đầu vào không hợp lệ"
                    },
                    "500": {
                        "description": "Lỗi máy chủ nội bộ"
                    }
                }
            }
        },
        "/api/auth/category/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Lấy danh mục theo ID",
                "operationId": "getOneCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy thành công danh mục"
                    },
                    "404": {
                        "description": "Không tìm thấy danh mục với ID"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Cập nhật thông tin danh mục",
                "description": "Cập nhật thông tin của danh mục bằng ID",
                "operationId": "a137bc8ac21cffdf6c1408c8c040000b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục cần cập nhật",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_name",
                                    "category_type"
                                ],
                                "properties": {
                                    "category_name": {
                                        "type": "string",
                                        "example": "Danh mục A"
                                    },
                                    "category_type": {
                                        "type": "string",
                                        "enum": [
                                            "product",
                                            "service",
                                            "post"
                                        ],
                                        "example": "product"
                                    },
                                    "slugs": {
                                        "type": "string",
                                        "example": "danh-muc-a"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Mô tả về danh mục A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật danh mục thành công"
                    },
                    "400": {
                        "description": "Yêu cầu không hợp lệ"
                    },
                    "404": {
                        "description": "Không tìm thấy danh mục với ID"
                    },
                    "500": {
                        "description": "Lỗi máy chủ"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Xóa danh mục",
                "description": "Xóa một danh mục theo ID",
                "operationId": "deleteCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục cần xóa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Xóa danh mục thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy danh mục với ID"
                    },
                    "500": {
                        "description": "Đã xảy ra lỗi trong quá trình xóa danh mục"
                    }
                }
            }
        },
        "/api/auth/category/{id}/force": {
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Xóa vĩnh viễn danh mục",
                "description": "Xóa một danh mục theo ID một cách vĩnh viễn",
                "operationId": "forceDeleteCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục cần xóa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Xóa danh mục thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy danh mục với ID"
                    },
                    "500": {
                        "description": "Đã xảy ra lỗi trong quá trình xóa danh mục"
                    }
                }
            }
        },
...
    },
...
}